+++ cursos-api/db/db.go
@@ -1,5 +1,41 @@
 package db
 
-func init() {
+import (
+	"context"
+	"cursos-api/model"
+	"fmt"
+	"log"
+	"os"
 
+	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func Init() {
+	if err := godotenv.Load(); err != nil {
+		log.Println("No .env file found")
+	}
+	var uri string
+	if uri = os.Getenv("MONGODB_URI"); uri == "" {
+		log.Fatal("You must set your 'MONGODB_URI' environment variable. See\n\t https://www.mongodb.com/docs/drivers/go/current/usage-examples/#environment-variable")
+	}
+	clientOptions := options.Client().ApplyURI(uri)
+	client, err := mongo.Connect(context.TODO(), clientOptions)
+	if err != nil {
+		panic(err)
+	}
+	defer func() {
+		if err = client.Disconnect(context.TODO()); err != nil {
+			panic(err)
+		}
+	}()
+
+	coll := client.Database("db").Collection("cursos")
+	newCourse := model.Curso{Nombre: "Rust", Precio: 14.21}
+	result, err := coll.InsertOne(context.TODO(), newCourse)
+	if err != nil {
+		panic(err)
+	}
+	fmt.Printf("Document inserted with ID: %s\n", result.InsertedID)
 }
