+++ cursos-api/main.go
@@ -1,38 +1,7 @@
 package main
 
-import (
-	"context"
-	"fmt"
-
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-)
-
-// Replace the placeholder with your Atlas connection string
-const uri = "mongodb://root:root@localhost:27017"
-
-type Curso struct {
-	Nombre string
-	Precio float64
-}
+import "cursos-api/db"
 
 func main() {
-	clientOptions := options.Client().ApplyURI("mongodb://root:root@localhost:27017")
-	client, err := mongo.Connect(context.TODO(), clientOptions)
-	if err != nil {
-		panic(err)
-	}
-	defer func() {
-		if err = client.Disconnect(context.TODO()); err != nil {
-			panic(err)
-		}
-	}()
-
-	coll := client.Database("db").Collection("cursos")
-	newCourse := Curso{Nombre: "8282", Precio: 14.21}
-	result, err := coll.InsertOne(context.TODO(), newCourse)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("Document inserted with ID: %s\n", result.InsertedID)
+	db.Init()
 }
