+++ cursos-api/model/curso.go
@@ -1,0 +1,6 @@
+package model
+
+type Curso struct {
+	Nombre string
+	Precio float32
+}
